requirements.txt
matplotlib==3.7.1
numpy==1.23.5
opencv_python==4.7.0.72
Pillow==9.3.0
Pillow==10.0.1
pyserial==3.5
scipy==1.11.3
tqdm==4.65.0
